+++ lib/pg/aws_rds_iam/auth_token_generator.rb
@@ -13,16 +13,39 @@ module PG
       def initialize(credentials:, region:)
         @generator = Aws::RDS::AuthTokenGenerator.new(credentials:)
         @region = region
+        @mutex = Mutex.new
+        @cache = {}
       end
 
       # Generates an authentication token for connecting to an Amazon RDS instance.
+      # Generated tokens are cached and reused until 1 minute before they are due to expire.
       #
       # @param host [String] the host name of the RDS instance that you want to access
       # @param port [String] the port number used for connecting to your RDS instance
       # @param user [String] the database account that you want to access
       # @return [String] the generated authentication token
       def call(host:, port:, user:)
-        @generator.auth_token(region: @region, endpoint: "#{host}:#{port}", user_name: user)
+        endpoint = "#{host}:#{port}"
+        key = "#{user}@#{endpoint}"
+
+        token = cached_token(key)
+        return token if token
+
+        @mutex.synchronize do
+          token = cached_token(key)
+          break token if token
+
+          @generator.auth_token(region: @region, endpoint:, user_name: user).tap do |new_token|
+            @cache[key] = AuthToken.new(new_token)
+          end
+        end
+      end
+
+      private
+
+      def cached_token(key)
+        token = @cache[key]
+        token.to_str if token&.valid?
       end
     end
   end